Update Link_Convertor_md_to_docx_pandoc_plugin.lua
Change error printing to warning(changed only the printed text) because there are cases where the link is not a link to a header or a note and then the filter does not work but not necessarily a error.

In case of warning, changed the return value from new link based on the input link to the original input link. The change has been made to support different types of links which my filter does not support.